Separação das funções UsufructSimulator.
